Reduce inconsistent results (#138)

+++ src/python/latex_output.py
@@ -78,7 +78,9 @@ def compute_trivial_merges(df: pd.DataFrame):
                         cache_merge_status_prefix + merge_tool + ".txt"
                     )
                     count += 1
-                    if DELETE_FAILED_TRIVIAL_MERGES:
+                    if DELETE_FAILED_TRIVIAL_MERGES and os.path.exists(
+                        cache_merges_status
+                    ):
                         os.remove(cache_merges_status)
                     else:
                         break
